refac: move _make_points to new utils file
